Update README to reflect addition of cron job
I added a cron job to automatically send a request every 30 minutes to the app as is hosted at https://kurushiid-stock-dashboard.herokuapp.com/ on Heroku, waking the app from sleep. The cron job will not execute between 12 am and 7 am, however; this is to fulfill the minimum time requirement for which a Heroku app must be asleep.